complete most features about settings, and fix something in LoginActivity, however, the sign-up feature is still to complete ( because of wrong cookie )
