+++ app/src/main/res/values-zh-rTW/strings.xml
@@ -15,4 +15,6 @@
     <string name="save_settings">保存更改</string>
     <string name="change_avatar">更換頭像</string>
     <string name="submit">提交</string>
+    <string name="password">密碼</string>
+    <string name="retype_password">驗證密碼</string>
 </resources>
