+++ app/src/main/res/values/styles.xml
@@ -22,9 +22,19 @@
         <item name="android:layout_marginLeft">10dp</item>
     </style>
 
+    <style name="TitleInSignUp">
+        <item name="android:layout_marginRight">10dp</item>
+    </style>
+
     <style name="QuestionItem">
         <item name="android:layout_marginLeft">10dp</item>
         <item name="android:layout_marginRight">10dp</item>
     </style>
 
+    <style name="EdtTxtInSignUp">
+        <item name="android:padding">10dp</item>
+        <item name="android:layout_marginTop">10dp</item>
+        <item name="android:layout_marginBottom">10dp</item>
+        <item name="android:background">@drawable/et</item>
+    </style>
 </resources>
